+++ src/main/java/com/redpxnda/respawnobelisks/registry/blocks/RespawnObeliskBlock.java
@@ -28,6 +28,10 @@ import net.minecraft.world.level.*;
 import net.minecraft.world.level.block.AnvilBlock;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.EntityBlock;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityTicker;
+import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.*;
@@ -44,9 +48,8 @@ import javax.annotation.Nullable;
 import java.util.List;
 import java.util.Optional;
 
-public class RespawnObeliskBlock extends Block {
+public class RespawnObeliskBlock extends Block implements EntityBlock {
     public static final EnumProperty<DoubleBlockHalf> HALF = BlockStateProperties.DOUBLE_BLOCK_HALF;
-    public static final IntegerProperty CHARGE = IntegerProperty.create("charge", 0, 3);
     public static final DirectionProperty RESPAWN_SIDE = DirectionProperty.create("respawn_side");
     private static final VoxelShape HITBOX_BOTTOM_BASE = Block.box(1.5D, 1.0D, 1.5D, 14.5D, 32.0D, 14.5D);
     private static final VoxelShape HITBOX_BOTTOM_TRIM = Block.box(0D, 0D, 0D, 16D, 3D, 16D);
@@ -59,7 +62,7 @@ public class RespawnObeliskBlock extends Block {
     public RespawnObeliskBlock(Properties pProperties, Either<ResourceKey<Level>, String> obeliskDimension) {
         super(pProperties);
         this.OBELISK_DIMENSION = obeliskDimension;
-        this.registerDefaultState(this.stateDefinition.any().setValue(HALF, DoubleBlockHalf.LOWER).setValue(CHARGE, 0).setValue(RESPAWN_SIDE, Direction.NORTH));
+        this.registerDefaultState(this.stateDefinition.any().setValue(HALF, DoubleBlockHalf.LOWER).setValue(RESPAWN_SIDE, Direction.NORTH));
     }
 
     public boolean propagatesSkylightDown(BlockState pState, BlockGetter pLevel, BlockPos pPos) {
@@ -98,51 +101,70 @@ public class RespawnObeliskBlock extends Block {
     }
 
     protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> pBuilder) {
-        pBuilder.add(HALF, CHARGE, RESPAWN_SIDE);
+        pBuilder.add(HALF, RESPAWN_SIDE);
+    }
+
+    @Override
+    public boolean hasAnalogOutputSignal(BlockState state) {
+        return state.getValue(HALF).equals(DoubleBlockHalf.LOWER);
+    }
+
+    @Override
+    public int getAnalogOutputSignal(BlockState state, Level level, BlockPos pos) {
+        if (state.getValue(HALF).equals(DoubleBlockHalf.LOWER) && level.getBlockEntity(pos) != null && level.getBlockEntity(pos) instanceof RespawnObeliskBlockEntity blockEntity)
+            return (int) (blockEntity.getCharge()*3f/20f);
+        return 0;
     }
 
     public Optional<Vec3> getRespawnLocation(BlockState state, BlockPos pos, ServerLevel level, ServerPlayer player) {
-        if (!(state.getValue(CHARGE) > 0 || player.hasEffect(Registry.IMMORTALITY_CURSE.get())))
-            if (ServerConfig.ENABLE_CURSE.get()) player.addEffect(new MobEffectInstance(Registry.IMMORTALITY_CURSE.get(), ServerConfig.CURSE_DURATION.get(), 0));
-        if (state.getValue(CHARGE) > 0) player.removeEffect(Registry.IMMORTALITY_CURSE.get());
-        Optional<Holder<Block>> blockHolder = ForgeRegistries.BLOCKS.getHolder(new ResourceLocation(ServerConfig.INFINITE_CHARGE_BLOCK.get()));
-        if (blockHolder.isPresent() && state.getValue(RespawnObeliskBlock.CHARGE) > 0) {
-            Class<? extends Block> blockClass = blockHolder.get().get().getClass();
-            boolean isNegative = !blockClass.isInstance(level.getBlockState(pos.below()).getBlock());
-            if (isNegative) {
-                level.setBlock(pos, state.setValue(CHARGE, state.getValue(CHARGE) - 1).setValue(HALF, DoubleBlockHalf.LOWER), 3);
-                level.setBlock(pos.above(), state.setValue(CHARGE, state.getValue(CHARGE) - 1).setValue(HALF, DoubleBlockHalf.UPPER), 3);
+        if (level.getBlockEntity(pos) != null && level.getBlockEntity(pos) instanceof RespawnObeliskBlockEntity blockEntity) {
+            double charge = blockEntity.getCharge();
+            if (!(charge > 0 || player.hasEffect(Registry.IMMORTALITY_CURSE.get())))
+                if (ServerConfig.ENABLE_CURSE.get())
+                    player.addEffect(new MobEffectInstance(Registry.IMMORTALITY_CURSE.get(), ServerConfig.CURSE_DURATION.get(), 0));
+            if (charge > 0) player.removeEffect(Registry.IMMORTALITY_CURSE.get());
+            Optional<Holder<Block>> blockHolder = ForgeRegistries.BLOCKS.getHolder(new ResourceLocation(ServerConfig.INFINITE_CHARGE_BLOCK.get()));
+            if (blockHolder.isPresent() && charge > 0) {
+                Class<? extends Block> blockClass = blockHolder.get().get().getClass();
+                boolean isNegative = !blockClass.isInstance(level.getBlockState(pos.below()).getBlock());
+                if (isNegative) {
+                    level.setBlockEntity(blockEntity.decreaseCharge(ServerConfig.OBELISK_DEPLETE_AMOUNT.get(), blockEntity.getLevel(), blockEntity.getBlockPos(), blockEntity.getBlockState()));
+                }
+                Packets.sendToPlayer(new RespawnObeliskSecondaryInteractionPacket(ServerConfig.OBELISK_DEPLETE_SOUND.get(), pos, isNegative, false), player);
+            } else {
+                if (ServerConfig.ENABLE_CURSE.get())
+                    Packets.sendToPlayer(new RespawnObeliskInteractionPacket(ServerConfig.CURSE_SOUND.get(), pos, true), player);
             }
-            Packets.sendToPlayer(new RespawnObeliskSecondaryInteractionPacket(ServerConfig.OBELISK_DEPLETE_SOUND.get(), pos, isNegative, false), player);
-        } else {
-            if (ServerConfig.ENABLE_CURSE.get()) Packets.sendToPlayer(new RespawnObeliskInteractionPacket(ServerConfig.CURSE_SOUND.get(), pos, true), player);
-        }
-        boolean isNotCurseEnabled = !ServerConfig.ENABLE_CURSE.get();
-        MobEffectInstance MEI = null;
-        if (player.hasEffect(Registry.IMMORTALITY_CURSE.get()))
-            MEI = player.getEffect(Registry.IMMORTALITY_CURSE.get());
-        boolean isMEI = MEI == null;
-        boolean isNotFullyCursed = ServerConfig.ENABLE_CURSE.get() && (isMEI || MEI.getAmplifier() < ServerConfig.MAX_CURSE_LEVEL.get()-1);
-        if (isNotCurseEnabled || isNotFullyCursed || state.getValue(CHARGE) > 0) {
-            BlockPos spawnPos = pos.relative(state.getValue(RESPAWN_SIDE));
-            return Optional.of(new Vec3(spawnPos.getX() + 0.5, spawnPos.getY(), spawnPos.getZ() + 0.5));
+            boolean isNotCurseEnabled = !ServerConfig.ENABLE_CURSE.get();
+            MobEffectInstance MEI = null;
+            if (player.hasEffect(Registry.IMMORTALITY_CURSE.get()))
+                MEI = player.getEffect(Registry.IMMORTALITY_CURSE.get());
+            boolean isMEI = MEI == null;
+            boolean isNotFullyCursed = ServerConfig.ENABLE_CURSE.get() && (isMEI || MEI.getAmplifier() < ServerConfig.MAX_CURSE_LEVEL.get() - 1);
+            if (isNotCurseEnabled || isNotFullyCursed || charge > 0) {
+                BlockPos spawnPos = pos.relative(state.getValue(RESPAWN_SIDE));
+                return Optional.of(new Vec3(spawnPos.getX() + 0.5, spawnPos.getY(), spawnPos.getZ() + 0.5));
+            }
         }
         return Optional.empty();
     }
+
+
 
     public InteractionResult use(BlockState pState, Level pLevel, BlockPos pPos, Player pPlayer, InteractionHand pHand, BlockHitResult pHit) {
         if (pLevel.isClientSide) {
             return InteractionResult.CONSUME;
         } else {
-            if (pPlayer instanceof ServerPlayer player) {
-                BlockPos finalPPos = pPos;
+            if (pState.getValue(HALF) == DoubleBlockHalf.UPPER) pPos = pPos.below();
+            if (pPlayer instanceof ServerPlayer player && pLevel.getBlockEntity(pPos) != null && pLevel.getBlockEntity(pPos) instanceof RespawnObeliskBlockEntity blockEntity) {
+                double charge = blockEntity.getCharge();
                 if (OBELISK_DIMENSION.left().isPresent()) {
                     if (pLevel.dimension() != OBELISK_DIMENSION.left().get()) {
-                        pLevel.setBlock(finalPPos, Blocks.AIR.defaultBlockState(), 3);
-                        pLevel.setBlock(finalPPos.above(), Blocks.AIR.defaultBlockState(), 3);
-                        pLevel.setBlock(finalPPos.below(), Blocks.AIR.defaultBlockState(), 3);
-                        pLevel.setBlock(finalPPos.below().below(), Blocks.AIR.defaultBlockState(), 3);
-                        pLevel.explode(null, DamageSource.badRespawnPointExplosion(), null, finalPPos.getX() + 0.5D, finalPPos.getY() + 0.5D, finalPPos.getZ() + 0.5D, 5.0F, true, Explosion.BlockInteraction.DESTROY);
+                        pLevel.setBlock(pPos, Blocks.AIR.defaultBlockState(), 3);
+                        pLevel.setBlock(pPos.above(), Blocks.AIR.defaultBlockState(), 3);
+                        pLevel.setBlock(pPos.below(), Blocks.AIR.defaultBlockState(), 3);
+                        pLevel.setBlock(pPos.below(2), Blocks.AIR.defaultBlockState(), 3);
+                        pLevel.explode(null, DamageSource.badRespawnPointExplosion(), null, pPos.getX() + 0.5D, pPos.getY() + 0.5D, pPos.getZ() + 0.5D, 5.0F, true, Explosion.BlockInteraction.DESTROY);
                         return InteractionResult.SUCCESS;
                     }
                 }
@@ -157,13 +179,13 @@ public class RespawnObeliskBlock extends Block {
                     int chargeAmount = Integer.parseInt(sections[1]); // 17 ^
                     boolean allowOverfill = sections.length >= 3 && Boolean.parseBoolean(sections[2]); // true ^
                     Optional<Holder<Item>> itemHolder = ForgeRegistries.ITEMS.getHolder(new ResourceLocation(sections[0]));
-                    boolean fullyCharged = pState.getValue(CHARGE) == 3 && chargeAmount > 0;
-                    boolean empty = pState.getValue(CHARGE) == 0 && chargeAmount < 0;
+                    boolean fullyCharged = charge == 100 && chargeAmount > 0;
+                    boolean empty = charge == 0 && chargeAmount < 0;
                     if (itemHolder.isPresent() &&
                             !player.getCooldowns().isOnCooldown(player.getMainHandItem().getItem()) &&
                             player.getMainHandItem().getItem() == itemHolder.get().get() // if held item = item
-                            && (pState.getValue(CHARGE) + chargeAmount <= 3 || allowOverfill) // if charge would go over 3, dont allow unless config says so
-                            && (pState.getValue(CHARGE) + chargeAmount >= 0 || allowOverfill) // if charge would go under 0, dont allow unless config says so
+                            && (charge + chargeAmount <= 100 || allowOverfill) // if charge would go over 100, dont allow unless config says so
+                            && (charge + chargeAmount >= 0 || allowOverfill) // if charge would go under 0, dont allow unless config says so
                             && !fullyCharged // if fully charged, don't allow more
                             && !empty // if empty charged, don't allow less
                     ) {
@@ -171,21 +193,8 @@ public class RespawnObeliskBlock extends Block {
                             Optional<Holder<Item>> itemHolder2 = ForgeRegistries.ITEMS.getHolder(new ResourceLocation(str2.split("\\|")[0]));
                             itemHolder2.ifPresent(holder -> player.getCooldowns().addCooldown(holder.get(), 30));
                         }
-                        Packets.sendToPlayer(new RespawnObeliskSecondaryInteractionPacket((chargeAmount < 0 ? ServerConfig.OBELISK_DEPLETE_SOUND.get() : ServerConfig.OBELISK_CHARGE_SOUND.get()), (pState.getValue(HALF) == DoubleBlockHalf.UPPER ? pPos.below() : pPos), (chargeAmount < 0), false), player);
-                        BlockState newState = pState.getValue(CHARGE) + chargeAmount >= 0 ?
-                                (pState.getValue(CHARGE) + chargeAmount <= 3 ? pState.setValue(CHARGE, pState.getValue(CHARGE) + chargeAmount) : pState.setValue(CHARGE, 3))
-                                : pState.setValue(CHARGE, 0);
-                        DoubleBlockHalf half;
-                        BlockPos offsetPos;
-                        if (pState.getValue(HALF) == DoubleBlockHalf.LOWER) {
-                            half = DoubleBlockHalf.UPPER;
-                            offsetPos = pPos.above();
-                        } else {
-                            half = DoubleBlockHalf.LOWER;
-                            offsetPos = pPos.below();
-                        }
-                        pLevel.setBlock(pPos, newState, 3);
-                        pLevel.setBlock(offsetPos, newState.setValue(HALF, half), 3);
+                        Packets.sendToPlayer(new RespawnObeliskSecondaryInteractionPacket((chargeAmount < 0 ? ServerConfig.OBELISK_DEPLETE_SOUND.get() : ServerConfig.OBELISK_CHARGE_SOUND.get()), pPos, (chargeAmount < 0), false), player);
+                        pLevel.setBlockEntity(blockEntity.increaseCharge(chargeAmount, blockEntity.getLevel(), blockEntity.getBlockPos(), blockEntity.getBlockState()));
                         if (!player.isCreative()) player.getMainHandItem().setCount(player.getMainHandItem().getCount()-1);
                         return InteractionResult.SUCCESS;
                     }
@@ -193,7 +202,6 @@ public class RespawnObeliskBlock extends Block {
                 // block removal
                 Optional<Holder<Item>> itemHolder = ForgeRegistries.ITEMS.getHolder(new ResourceLocation(ServerConfig.REMOVAL_ITEM.get()));
                 if (ServerConfig.ALLOW_PICKUP.get() && itemHolder.isPresent() && player.getMainHandItem().getItem() == itemHolder.get().get()) {
-                    if (pState.getValue(HALF) == DoubleBlockHalf.UPPER) pPos = pPos.below();
                     pLevel.setBlock(pPos, Blocks.AIR.defaultBlockState(), 3);
                     pLevel.setBlock(pPos.above(), Blocks.AIR.defaultBlockState(), 3);
                     if (!player.isCreative()) player.getMainHandItem().setCount(player.getMainHandItem().getCount()-1);
@@ -214,7 +222,7 @@ public class RespawnObeliskBlock extends Block {
                     }
                     return InteractionResult.SUCCESS;
                 // checking if there's no charge
-                } else if (pState.getValue(CHARGE) == 0) {
+                } else if (charge == 0) {
                     player.sendSystemMessage(Component.translatable("text.respawnobelisks.no_charge"));
                 } else {
                     int degrees;
@@ -222,7 +230,6 @@ public class RespawnObeliskBlock extends Block {
                     else if (pState.getValue(RESPAWN_SIDE) == Direction.EAST) degrees = -90;
                     else if (pState.getValue(RESPAWN_SIDE) == Direction.SOUTH) degrees = 0;
                     else degrees = 90;
-                    if (!(pState.getValue(HALF) == DoubleBlockHalf.LOWER)) pPos = pPos.below();
                     if ((player.getRespawnPosition() != null && !player.getRespawnPosition().equals(pPos)) || player.getRespawnPosition() == null)
                         Packets.sendToPlayer(new RespawnObeliskSecondaryInteractionPacket(ServerConfig.OBELISK_SET_RESPAWN_SOUND.get(), pPos, false, true), player);
                     player.setRespawnPosition(pLevel.dimension(), pPos, degrees, false, true);
@@ -231,5 +238,12 @@ public class RespawnObeliskBlock extends Block {
             }
         }
         return InteractionResult.FAIL;
+    }
+
+    @Nullable
+    @Override
+    public BlockEntity newBlockEntity(BlockPos pPos, BlockState pState) {
+        if (pState.getValue(HALF).equals(DoubleBlockHalf.UPPER)) return null;
+        return Registry.RESPAWN_OBELISK_BE.get().create(pPos, pState);
     }
 }